--- backend/src/main/java/ddangkong/domain/balance/room/RoomStatus.java
@@ -1,20 +1,0 @@
-package ddangkong.domain.balance.room;
-
-public enum RoomStatus {
-    READY,
-    PROGRESS,
-    FINISH,
-    ;
-
-    public boolean isAlreadyStart() {
-        return this != READY;
-    }
-
-    public boolean isGameProgress() {
-        return this == PROGRESS;
-    }
-
-    public boolean isGameFinish() {
-        return this == FINISH;
-    }
-}
