+++ backend/src/main/java/ddangkong/domain/room/RoomStatus.java
@@ -1,0 +1,20 @@
+package ddangkong.domain.room;
+
+public enum RoomStatus {
+    READY,
+    PROGRESS,
+    FINISH,
+    ;
+
+    public boolean isAlreadyStart() {
+        return this != READY;
+    }
+
+    public boolean isGameProgress() {
+        return this == PROGRESS;
+    }
+
+    public boolean isGameFinish() {
+        return this == FINISH;
+    }
+}
