+++ backend/src/main/java/ddangkong/service/balance/content/dto/BalanceContentResponse.java
@@ -1,9 +1,9 @@
 package ddangkong.service.balance.content.dto;
 
-import ddangkong.service.balance.option.dto.BalanceOptionResponse;
 import ddangkong.domain.balance.content.Category;
 import ddangkong.domain.balance.option.BalanceOptions;
-import ddangkong.domain.balance.room.RoomContent;
+import ddangkong.domain.room.balance.roomcontent.RoomContent;
+import ddangkong.service.balance.option.dto.BalanceOptionResponse;
 import lombok.Builder;
 
 public record BalanceContentResponse(
