--- backend/src/main/java/ddangkong/service/balance/member/dto/MemberResponse.java
@@ -1,13 +1,0 @@
-package ddangkong.service.balance.member.dto;
-
-import ddangkong.domain.member.Member;
-
-public record MemberResponse(
-        Long memberId,
-        String nickname,
-        boolean isMaster
-) {
-    public MemberResponse(Member member) {
-        this(member.getId(), member.getNickname(), member.isMaster());
-    }
-}
