--- backend/src/main/java/ddangkong/service/balance/room/dto/RoomJoinRequest.java
@@ -1,9 +1,0 @@
-package ddangkong.service.balance.room.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RoomJoinRequest(
-        @NotBlank
-        String nickname
-) {
-}
