--- backend/src/main/java/ddangkong/service/balance/room/dto/RoomJoinResponse.java
@@ -1,9 +1,0 @@
-package ddangkong.service.balance.room.dto;
-
-import ddangkong.service.balance.member.dto.MemberResponse;
-
-public record RoomJoinResponse(
-        Long roomId,
-        MemberResponse member
-) {
-}
