+++ backend/src/main/java/ddangkong/service/balance/vote/BalanceVoteService.java
@@ -5,14 +5,14 @@ import ddangkong.domain.balance.content.BalanceContentRepository;
 import ddangkong.domain.balance.option.BalanceOption;
 import ddangkong.domain.balance.option.BalanceOptionRepository;
 import ddangkong.domain.balance.option.BalanceOptions;
-import ddangkong.domain.balance.room.Room;
-import ddangkong.domain.balance.room.RoomContent;
-import ddangkong.domain.balance.room.RoomContentRepository;
-import ddangkong.domain.balance.room.RoomRepository;
 import ddangkong.domain.balance.vote.BalanceVote;
 import ddangkong.domain.balance.vote.BalanceVoteRepository;
-import ddangkong.domain.member.Member;
-import ddangkong.domain.member.MemberRepository;
+import ddangkong.domain.room.Room;
+import ddangkong.domain.room.RoomRepository;
+import ddangkong.domain.room.balance.roomcontent.RoomContent;
+import ddangkong.domain.room.balance.roomcontent.RoomContentRepository;
+import ddangkong.domain.room.member.Member;
+import ddangkong.domain.room.member.MemberRepository;
 import ddangkong.exception.BadRequestException;
 import ddangkong.service.balance.content.dto.BalanceContentGroupResponse;
 import ddangkong.service.balance.content.dto.BalanceContentTotalResponse;
