+++ backend/src/main/java/ddangkong/service/room/dto/RoomJoinRequest.java
@@ -1,0 +1,9 @@
+package ddangkong.service.room.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RoomJoinRequest(
+        @NotBlank
+        String nickname
+) {
+}
