+++ backend/src/main/java/ddangkong/service/room/dto/RoomJoinResponse.java
@@ -1,0 +1,9 @@
+package ddangkong.service.room.dto;
+
+import ddangkong.service.room.member.dto.MemberResponse;
+
+public record RoomJoinResponse(
+        Long roomId,
+        MemberResponse member
+) {
+}
