+++ backend/src/main/java/ddangkong/service/room/dto/RoundFinishedResponse.java
@@ -1,0 +1,7 @@
+package ddangkong.service.room.dto;
+
+public record RoundFinishedResponse(
+        boolean isRoundFinished,
+        boolean isGameFinished
+) {
+}
