+++ backend/src/test/java/ddangkong/controller/balance/vote/BalanceVoteControllerTest.java
@@ -10,10 +10,10 @@ import ddangkong.controller.BaseControllerTest;
 import ddangkong.domain.balance.content.BalanceContent;
 import ddangkong.domain.balance.content.Category;
 import ddangkong.domain.balance.option.BalanceOption;
-import ddangkong.domain.balance.room.Room;
-import ddangkong.domain.balance.room.RoomContent;
 import ddangkong.domain.balance.vote.BalanceVote;
-import ddangkong.domain.member.Member;
+import ddangkong.domain.room.Room;
+import ddangkong.domain.room.balance.roomcontent.RoomContent;
+import ddangkong.domain.room.member.Member;
 import ddangkong.service.balance.vote.dto.BalanceVoteRequest;
 import ddangkong.service.balance.vote.dto.BalanceVoteResponse;
 import ddangkong.service.balance.vote.dto.VoteFinishedResponse;
