+++ backend/src/test/java/ddangkong/service/balance/vote/BalanceVoteServiceTest.java
@@ -10,10 +10,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import ddangkong.domain.balance.content.BalanceContent;
 import ddangkong.domain.balance.content.Category;
 import ddangkong.domain.balance.option.BalanceOption;
-import ddangkong.domain.balance.room.Room;
-import ddangkong.domain.balance.room.RoomContent;
 import ddangkong.domain.balance.vote.BalanceVote;
-import ddangkong.domain.member.Member;
+import ddangkong.domain.room.Room;
+import ddangkong.domain.room.balance.roomcontent.RoomContent;
+import ddangkong.domain.room.member.Member;
 import ddangkong.exception.BadRequestException;
 import ddangkong.service.BaseServiceTest;
 import ddangkong.service.balance.content.dto.BalanceContentGroupResponse;
