+++ backend/src/test/java/ddangkong/support/fixture/MemberFixture.java
@@ -1,7 +1,7 @@
 package ddangkong.support.fixture;
 
-import ddangkong.domain.balance.room.Room;
-import ddangkong.domain.member.Member;
+import ddangkong.domain.room.Room;
+import ddangkong.domain.room.member.Member;
 
 public enum MemberFixture {
     PRIN,
